add pos struct Signature json tag
